fix nell'esecuzione della simulazione fix time_format hgignore aggiornato rimossi i config_b0{2,3}.conf deprecati

+++ src/common/timer/time.cpp
@@ -20,7 +20,7 @@ void time_stop(timer &time)
     gettimeofday(&time.stop,NULL);
 }
 
-string time_format(uint32 sec)
+string time_format(float sec)
 {
     stringstream out;
